move engine to an isolated namespace
This solves the problem of routes collision with the main app. For
instance where there is a catch-all routes like `get ':slug', to:
'pages#show'`, without the isolated namespace, the engine's routes are
appended to the main application's routes file, but they should be
prepended to work. After this change the engine's routes can be inserted
anywhere in the main application's routes file.